[v0.90.2][runtime] Split governed episode records (Closes #2309)
